Adjust docker script when using podman
In podman, a user namespace is setup with IDs mapped such that the
executing user is root in the container. When that's the case (or if
docker is in use and the executing user is root), don't override the
container user or the HOME environment variable. In that case, the
container user's home directory will be /root, so mount the cache there
instead of the predefined /cache.