Makefile: Do not remove apps-bundle as part of clean
Create a custom rule to remove clear the apps-bundle and use this one in
the `src/apps-bundle` instead of clean.

This prevents the issue of removing the apps-bundle after creating it
and other future possible issues of removing `src/kolibri` or other data
when the user just wants to rebuild the apps-bundle.

https://phabricator.endlessm.com/T33466