Add Toolbox image for development
The documented development flow doesn't work great on Endless because
you can't install Java. Similarly, the documented docker/podman flow is
not friendly for development since you don't have your full development
environment.

This commit adds a second image to be used with toolbox. It just
contains Java and the Android SDK. The rest can be done in a shell in
the toolbox container like you would for any other python project.