Vendor DynamicWhiteNoise from Kolibri
In order to access files via DocumentsProvider, the DynamicWhiteNoise
static files class in Kolibri needs to be adjusted. Since Kolibri
doesn't provide a pluggable implementation, it needs to be monkey
patched in. Unfortunately, that means it can't be subclassed since
Python goes into infinite recursion when DynamicWhiteNoise tries to
chain up to its parent class.

https://phabricator.endlessm.com/T33662