Add QEMU packages to toolbox image
The Android emulator in the SDK does not contain all the host
dependencies as these are expected to ship in Android Studio (I
believe). Install the QEMU packages from Endless as a means to get the
them. Also include qemu-utils for `qemu-img` and `qemu-nbd` even though
`qemu-nbd` won't work unless you're real root in the container.

https://phabricator.endlessm.com/T33535