Skip Chrome OS My Files removable volume
On Chrome OS, the My Files volume is marked as removable with Android.
That throws off detection of true removable drives when trying to
determine if the Endless Key USB has been inserted. It turns out this
volume has a constant UUID in Chrome OS, so use that to filter it out.

https://phabricator.endlessm.com/T33569