Clean p4a cached builds in Jenkins
Unfortunately, there are just too many cache invalidation bugs in p4a to
know if it's done the right thing or not. In addition to the already
deleted distributions, delete the builds as well. All that's left then
is the cached package downloads, which I believe to be safe. This will
make the builds more robust at the expense of making take a lot longer.

https://phabricator.endlessm.com/T33880