Restore normal ccache usage
The zstandard build was the only one know to be affected by p4a's use of
`ccache`, and that's been fixed. While the method of putting
`/usr/lib/ccache` in `PATH` works fine for the host compiler where there
are symlinks in the directory, that actually doesn't help much in the
python-for-android scenario. In that case, most of the compiling happens
using the NDK's cross compilers for all the target architectures. Those
are definitely not installed in `/usr/lib/ccache` and can only be used
by prepending the command with `ccache`.

https://phabricator.endlessm.com/T33880