Rewrite Dockerfile to more closely resemble toolbox image
This fixes an issue where newer p4a requires python 3.7, which is not
available in Ubuntu bionic. However, it makes the build environment used
in Jenkins more closely resemble the toolbox based one that developers
use. That makes it more likely that the published build matches what
developers test locally.

Some of the system packages needed for building the app were only part
of the toolbox image as part of the eos-toolbox-dev metapackage. Those
have been added explicitly now.

https://phabricator.endlessm.com/T33884