Set CONTENT_FALLBACK_DIRS after USB URIs set
When Kolibri was running from a separate service, setting
`CONTENT_FALLBACK_DIRS` early in `initialization.py` was fine since the
service process wouldn't be spawned until after the initial USB checks.
Now Kolibri is running in the main activity, so the environment variable
needs to be set immediately after the USB checks, too.

https://phabricator.endlessm.com/T33793