README: Add a note for p4a commnad not found
The python-for-android providing "p4a" command is installed from the
requirements.txt in this repository. The installation is bounded in the
user enviroment by default, not global.

However, not every distribution's environment variable "PATH" includes
the user's HOME path "~/.local/bin". That leads the "p4a" command in
"~/.local/bin" becoming missed. So, add this note directing to the
solution: "p4a: command not found" [1].

[1]: https://github.com/kivy/python-for-android/issues/1167