Remove duplicate code from android_whitenoise
By adding a patch to lightly refactor kolibri_whitenoise, we can change
android_whitenoise to use a subclass of DynamicWhiteNoise.

It is still necessary to copy DynamicWhiteNoise in order to subclass it,
because the same class in Kolibri itself needs to be monkey-patched with
our version. However, that copy can now be created automatically.

https://phabricator.endlessm.com/T34559