Add custom Android logging handler
Currently the way python logging records end up in the Android log is
that Kolibri provides handlers sending records to stdout and stderr and
then p4a takes lines sent to stdout and stderr and converts them to Log
messages.

This works but it's suboptimal. It's inefficient to have the log record
converted into a stream so p4a can split the stream back up, p4a logs
everything at verbose priority since it doesn't know the log level, and
p4a uses the extremely generic `python` tag for everything.

Using a logging handler allows sending the log directly to the Android
log system while preserving the format and level of the record.